// src/models/rsi-divergence-types.enum.ts
export enum RsiDivergenceTypes {
  Bullish = "RSI Bullish",
  Bearish = "RSI Bearish",
  HiddenBullish = "RSI Hidden Bullish",
  HiddenBearish = "RSI Hidden Bearish",
  NotAvailable = "NO RSI Divergence",
}

export enum RsiDivergenceDirection {
  Bullish = "bullish",
  Bearish = "bearish",
  NotAvailable = "notAvailable",
}